Move link resolvers to separate lib
