+++ client/src/lib/linking/utils/url.spec.ts
@@ -1,4 +1,4 @@
-import {appendOrigin} from './utils';
+import {appendOrigin} from './url';
 
 describe('appendOrigin', () => {
   it('adds origin search param to url', () => {
